backend/pyproject.toml
[project]
name = "sheepstack-backend"
version = "0.1.0"
description = "SheepStack v3.0 Backend - FastAPI + LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # Database
    "sqlalchemy>=2.0.25",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    # LangChain & LangGraph
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-postgres>=0.0.1",
    "langchain-openai>=0.0.5",
    # CrewAI
    "crewai>=0.1.0",
    # OpenTelemetry
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    # Supabase
    "supabase>=2.3.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "pypdf>=6.4.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.0.0",
    "crewai-tools>=1.6.0",
    "litellm>=1.80.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "faker>=22.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=xml",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "src/main.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
